requirements.txt
setuptools<82
Django==4.1.7
django-bootstrap-v5==1.0.11
django-polymorphic==3.1.0
django-template-maths==0.2.0
numpy
plotly
django-ajax-datatable
